# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JABEEN Investor Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/updates:
    get:
      operationId: listUpdates
      tags: [updates]
      summary: List project status updates
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Update list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusUpdate"
    post:
      operationId: createUpdate
      tags: [updates]
      summary: Submit a status update
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInput"
      responses:
        "201":
          description: Created update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusUpdate"
        "409":
          description: Pending update already exists or approval guard failed
  /projects/{projectId}/updates/{updateId}/approve:
    patch:
      operationId: approveUpdate
      tags: [updates]
      summary: Approve a pending update (managers only)
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/updateId"
      responses:
        "200":
          description: Approved update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusUpdate"
        "409":
          description: Approval guard failed
  /projects/{projectId}/updates/{updateId}/reject:
    patch:
      operationId: rejectUpdate
      tags: [updates]
      summary: Reject a pending update (managers only)
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/updateId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusUpdate"

  # --- DOCUMENTS ---
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List stage templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemplateSummary"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a stage template (managers only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{templateId}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get template detail
      parameters:
        - $ref: "#/components/parameters/templateId"
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    put:
      operationId: replaceTemplate
      tags: [templates]
      summary: Replace template stages/fields (managers only)
      parameters:
        - $ref: "#/components/parameters/templateId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template (managers only)
      parameters:
        - $ref: "#/components/parameters/templateId"
      responses:
        "204":
          description: Deleted

  # --- USERS ---
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
    updateId:
      name: updateId
      in: path
      required: true
      schema:
        type: integer
...
    templateId:
      name: templateId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    User:
      type: object
      required: [id, email, fullName, companyName, role, active, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        companyName:
          type: string
        title:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [investor, top-management, project-manager, administrator]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    StageSummary:
      type: object
      required: [id, name, orderIndex, progressBaseline, category]
      properties:
        id:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer
        progressBaseline:
          type: integer
        category:
          type: string
          enum: [on-hold, active, complete]
        description:
          type: ["string", "null"]
...
    StatusUpdate:
      type: object
      required: [id, projectId, authorId, targetStageId, constructionPct, reviewStatus, fieldValues, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        sourceStageId:
          type: ["integer", "null"]
        targetStageId:
          type: integer
        targetStage:
          $ref: "#/components/schemas/StageSummary"
        constructionPct:
          type: integer
        note:
          type: ["string", "null"]
        reviewStatus:
          type: string
          enum: [pending, approved, rejected]
        reviewerId:
          type: ["integer", "null"]
        reviewer:
          $ref: "#/components/schemas/User"
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        reviewNote:
          type: ["string", "null"]
        documentId:
          type: ["integer", "null"]
        fieldValues:
          type: array
          items:
            $ref: "#/components/schemas/FieldValue"
        createdAt:
          type: string
          format: date-time
    UpdateInput:
      type: object
      required: [targetStageId]
      properties:
        targetStageId:
          type: integer
        constructionPct:
          type: integer
          minimum: 0
          maximum: 100
        note:
          type: string
          maxLength: 4000
        fieldValues:
          type: array
          items:
            $ref: "#/components/schemas/FieldValueInput"
    RejectInput:
      type: object
      properties:
        reviewNote:
          type: string
          maxLength: 4000
    FieldValue:
      type: object
      required: [id, fieldId]
      properties:
        id:
          type: integer
        fieldId:
          type: integer
        fieldName:
          type: string
        textValue:
          type: ["string", "null"]
        numValue:
          type: ["number", "null"]
        dateValue:
          type: ["string", "null"]
        boolValue:
          type: ["boolean", "null"]
        choiceValue:
          type: ["string", "null"]
    FieldValueInput:
      type: object
      required: [fieldId]
      properties:
        fieldId:
          type: integer
        textValue:
          type: string
        numValue:
          type: number
        dateValue:
          type: string
        boolValue:
          type: boolean
        choiceValue:
          type: string

    # --- DOCUMENT ---
...
    TemplateSummary:
      type: object
      required: [id, name, isDefault, stageCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isDefault:
          type: boolean
        stageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    Template:
      type: object
      required: [id, name, isDefault, stages, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isDefault:
          type: boolean
        stages:
          type: array
          items:
            $ref: "#/components/schemas/Stage"
        createdAt:
          type: string
          format: date-time
    TemplateInput:
      type: object
      required: [name, stages]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 500
        isDefault:
          type: boolean
        stages:
          type: array
          items:
            $ref: "#/components/schemas/StageInput"
    Stage:
      type: object
      required: [id, name, orderIndex, progressBaseline, category, fields]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        orderIndex:
          type: integer
        progressBaseline:
          type: integer
          minimum: 0
          maximum: 100
        category:
          type: string
          enum: [on-hold, active, complete]
        fields:
          type: array
          items:
            $ref: "#/components/schemas/StageField"
    StageInput:
      type: object
      required: [name, progressBaseline, category]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 500
        progressBaseline:
          type: integer
          minimum: 0
          maximum: 100
        category:
          type: string
          enum: [on-hold, active, complete]
        fields:
          type: array
          items:
            $ref: "#/components/schemas/StageFieldInput"
    StageField:
      type: object
      required: [id, name, baseType, widget, required, position]
      properties:
        id:
          type: integer
        name:
          type: string
        baseType:
          type: string
          enum: [text, number, date, boolean, file, image, single-choice, multi-choice]
        widget:
          type: string
          enum: [single-line, multi-line, email, telephone, number, date, checkbox, toggle, file-upload, single-photo, photo-gallery, drop-list, list-box, radio, checkbox-list]
        required:
          type: boolean
        position:
          type: integer
        options:
          type: array
          items:
            type: string
        config:
          type: object
    StageFieldInput:
      type: object
      required: [name, baseType, widget, required]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        baseType:
          type: string
          enum: [text, number, date, boolean, file, image, single-choice, multi-choice]
        widget:
          type: string
          enum: [single-line, multi-line, email, telephone, number, date, checkbox, toggle, file-upload, single-photo, photo-gallery, drop-list, list-box, radio, checkbox-list]
        required:
          type: boolean
        options:
          type: array
          items:
            type: string
        config:
          type: object

    # --- AUDIT ---
...